+++ src/Message/MessageHandler.php
@@ -13,7 +13,6 @@ use Psr\Log\{
 };
 use Stringable;
 use WebSocket\Configuration;
-use WebSocket\Exception\BadOpcodeException;
 use WebSocket\Frame\{
     Frame,
     FrameHandler,
@@ -65,7 +64,7 @@ class MessageHandler implements LoggerAwareInterface, Stringable
      */
     public function push(Message $message, int $size = self::DEFAULT_SIZE): Message
     {
-        $frames = $message->getFrames($size);
+        $frames = $message->getFrames($size, $this->configuration->getOpcodeRegistry());
         foreach ($frames as $frame) {
             $this->frameHandler->push($frame);
         }
@@ -100,19 +99,13 @@ class MessageHandler implements LoggerAwareInterface, Stringable
 
     /**
      * @param non-empty-array<Frame> $frames
-     * @throws BadOpcodeException
      */
     private function createMessage(array $frames): Message
     {
-        $opcode = $frames[0]->getOpcode() ?? null;
-        $message = match ($opcode) {
-            'text' => new Text(),
-            'binary' => new Binary(),
-            'ping' => new Ping(),
-            'pong' => new Pong(),
-            'close' => new Close(),
-            default => throw new BadOpcodeException("Invalid opcode '{$opcode}' provided"),
-        };
+        /** @var int<1, 15> $opcode */
+        $opcode = $frames[0]->getOpcode();
+        $message = $this->configuration->getOpcodeRegistry()->createMessage($opcode);
+
         $message->setPayload(array_reduce($frames, function (string $carry, Frame $item) {
             return $carry . $item->getPayload();
         }, ''));