+++ tests/suites/message/BinaryTest.php
@@ -14,7 +14,7 @@ use Stringable;
 use WebSocket\Frame\Frame;
 use WebSocket\Message\{
     Binary,
-    Message
+    Message,
 };
 
 /**